Destructure sections on the fly.
Now there is no need to deal with section-text and section-end-TeX.
(In fact, those slots are completely gone now.) The three parts of
a section are split up into the slots TeX-part, definition-part,
and Pascal-part (all lists of tokens), whose use should be
intuitive. Macro definitions are still not handled, but adding them
would not be as awkward as before.

Also, the definitions slot of the unnamed module was not nreversed
in the previous version; this has been fixed.