Fix scope problem with WEAVE
WEAVE defines a numeric macro called |close|, which has the same
name as a predefined Pascal-H procedure. Previously, macros were
installed before standard identifiers, so that the procedure
|close| would shadow the macro. This lead to a problem in the |case|
statement in |print_cat|, since the parser expected |close| to be
given an argument.

I fixed this for my statistic gathering (https://github.com/shreevatsa/webWEB/discussions/15)
simply by temporarily removing the installation of the |close|
procedure, but it is far better to change the order of operations:
Now standard identifiers are installed *before* macros; thus the
|close| macro takes precedent.

This also means that TeX will work without removing the definition
of |text|, but only because there are no program parameter files.
(It would be more robust to store the underlying type of |text|,
rather than to rely on the existence of some type by that name.
However, the latter option requires deciding where to put the
representation, and I don't know what the best place would be.)