Add the |only-meaning| slot and |map-identifiers|
This is in preparation for inline code handling, which is
necessarily heuristic. If we see an identifier, and it has only
one meaning, then that meaning must be what the identifier is
referring to. I wanted to see how effective this heuristic would
be, and the results are promising: There are ten times as many
identifiers with only one meaning than identifiers with multiple
meanings (or with no meaning).

The |map-identifiers| function was defined for testing purposes,
but it is useful to have in general.